Move AutoArmy out of CPlayer
commit-id:d06c6cae
