+++ MekWarsClient/src/mekwars/client/campaign/CArmy.java
@@ -41,7 +41,7 @@ public class CArmy extends Army<CUnit> {
      * from CPlayer.setArmyData() and CPlayer.setData(). In both instances, the string being passed
      * is generated by the server and sent down using a PL| cmd.
      */
-    public void fromString(String s, CPlayer p, String delimiter, MWClient client) {
+    public void fromString(String s, CPlayer p, String delimiter) {
         StringTokenizer ST = new StringTokenizer(s, delimiter);
         setId(TokenReader.readInt(ST));
         setBV(TokenReader.readInt(ST));
